+++ packages/blocks/core/js/wordpress/post-navigation-link/index.js
@@ -1,9 +1,15 @@
 // @flow
 
 /**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Blockera dependencies
  */
 import { SharedBlockExtension } from '@blockera/editor';
+import { Icon } from '@blockera/icons';
 
 /**
  * Internal dependencies
@@ -15,7 +21,25 @@ export const PostNavigationLink: BlockType = {
 	name: 'blockeraPostNavigationLink',
 	targetBlock: 'core/post-navigation-link',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/arrow': {
+			name: 'elements/arrow',
+			type: 'title',
+			label: __('Arrow', 'blockera'),
+			description: __('Arrow icon element.', 'blockera'),
+			icon: <Icon icon="previous" library="wp" iconSize="20" />,
+			settings: {
+				force: true,
+			},
+		},
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			label: __('Link', 'blockera'),
+			description: __('Hyperlink element.', 'blockera'),
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;